+++ tests/Unit/QueryTest.php
@@ -364,4 +364,32 @@ class QueryTest extends \PHPUnit_Framework_TestCase
 
         $this->assertEquals('limit=150&skip=10&content_type=cat&order=sys.createdAt&sys.id=nyancat&sys.updatedAt%5Blte%5D=2013-01-01T00%3A00%3A00%2B00%3A00', (string) $queryBuilder->getQueryString());
     }
+
+    /**
+     * @covers Contentful\Query::__construct
+     * @covers Contentful\Query::select
+     * @covers Contentful\Query::setContentType
+     * @covers Contentful\Query::getQueryData
+     * @covers Contentful\Query::getQueryString
+     */
+    public function testQueryWithSelect()
+    {
+        $queryBuilder = (new Query)
+            ->select(['foobar1'])
+            ->setContentType('cat');
+
+        $this->assertEquals('content_type=cat&select=sys%2Cfoobar1', $queryBuilder->getQueryString());
+
+        $queryBuilder = (new Query)
+            ->select(['foobar2'])
+            ->setContentType('cat');
+
+        $this->assertEquals('content_type=cat&select=sys%2Cfoobar2', $queryBuilder->getQueryString());
+
+        $queryBuilder = (new Query)
+            ->select(['sys'])
+            ->setContentType('cat');
+
+        $this->assertEquals('content_type=cat&select=sys', $queryBuilder->getQueryString());
+    }
 }